додає стилі секції Subscribe
