+++ src/partials/customers.html
@@ -1,1 +1,1 @@
-<section class="section customers">Customers</section>
+<section>Customers</section>
