+++ src/sass/layout/_customers.scss
@@ -1,4 +1,1 @@
-.customers {
-  margin-bottom: 52px;
-  min-height: 769px;
-}
+
