+++ src/sass/layout/_subscribe.scss
@@ -1,23 +1,64 @@
 // Subscribe
 
 .subscribe__container {
+  padding-top: 52px;
+  padding-bottom: 60px;
   position: relative;
+
+  @media screen and (max-width: 767.98px) {
+    @include centeredFlex(flex, center, center);
+    flex-direction: column;
+  }
+
+  @media screen and (min-width: $tablet) {
+    padding-top: 82px;
+    padding-bottom: 100px;
+  }
+
+  @media screen and (min-width: $desktop) {
+    padding-top: 0;
+  }
 }
 
 .subscribe__img {
   position: absolute;
-  top: -95px;
+
+  @media screen and (max-width: 767.98px) {
+    top: -43px;
+  }
+
+  @media screen and (min-width: $tablet) {
+    top: -74px;
+    right: 32px;
+  }
+
+  @media screen and (min-width: $desktop) {
+    top: -66px;
+    right: 16px;
+  }
+}
+
+.subscribe__description {
+  @media screen and (min-width: $desktop) {
+    width: 656px;
+  }
 }
 
 .subscribe__text {
   margin-bottom: 40px;
 
   @include fonts(14px, 16px, 400);
+
+  @media screen and (min-width: $tablet) {
+    margin-bottom: 24px;
+    @include fonts(16px, 24px, 400);
+  }
 }
 
 //Subscribe Form
 
 .subscription-form__field {
+  position: relative;
   @media screen and (max-width: 767.98px) {
     @include centeredFlex(flex, flex-end, center);
 
@@ -28,25 +69,65 @@
 
 .subscription-form__input {
   width: 100%;
-  padding: 12px 11px;
+  padding: 12px;
 
   color: $color-text-primary;
   @include fonts(12px, 18px, 400);
 
   border: 1px solid #ffe550;
   border-radius: 30px;
+  outline: transparent;
+
+  @media screen and (min-width: $tablet) {
+    margin-right: 22px;
+    max-width: 395px;
+  }
+
+  @media screen and (min-width: $desktop) {
+    margin-right: 32px;
+  }
 
   &::placeholder {
     color: $color-form-placeholder;
     @include fonts(12px, 18px, 400);
   }
+
+  &:valid:not(:placeholder-shown) {
+    border-color: $color-btn-border;
+  }
+
+  &:invalid:not(:placeholder-shown) {
+    border-color: $color-form-error;
+  }
+
+  &:focus-within {
+    border-color: $color-btn-border-hover;
+  }
+}
+
+.subscription-form__error {
+  position: absolute;
+  top: 48px;
+  left: 0;
+
+  @include centeredFlex();
+  display: none;
+  gap: 4px;
+  color: $color-form-error;
+  @include fonts(10px, 15px, 400);
+}
+
+.subscription-form__input:invalid:not(:placeholder-shown)
+  + .subscription-form__error {
+  display: block;
 }
 
 .subscription-form__btn {
+  padding: 11px 22px;
   color: $color-text-primary;
   background-color: $color-accent;
   border-radius: 30px;
-  border-color: transparent;
+  border: none;
 
   &:hover,
   &:focus {
